checkShipPlacement validation on server side
